+++ http/sentryhttp_test.go
@@ -216,6 +216,7 @@ func TestIntegration(t *testing.T) {
 			sentry.Event{},
 			"Contexts", "EventID", "Extra", "Platform", "Modules",
 			"Release", "Sdk", "ServerName", "Tags", "Timestamp",
+			"sdkMetaData",
 		),
 		cmpopts.IgnoreFields(
 			sentry.Request{},