+++ sentry.go
@@ -5,15 +5,21 @@ import (
 	"time"
 )
 
+// Deprecated: Use SDKVersion instead.
+const Version = SDKVersion
+
 // Version is the version of the SDK.
-const Version = "0.15.0"
+const SDKVersion = "0.16.0"
+
+// The identifier of the SDK.
+const SDKIdentifier = "sentry.go"
 
 // apiVersion is the minimum version of the Sentry API compatible with the
 // sentry-go SDK.
 const apiVersion = "7"
 
 // userAgent is the User-Agent of outgoing HTTP requests.
-const userAgent = "sentry-go/" + Version
+const userAgent = "sentry-go/" + SDKVersion
 
 // Init initializes the SDK with options. The returned error is non-nil if
 // options is invalid, for instance if a malformed DSN is provided.